Update Ray.py
added a cone feature,
This enables the 3d ray beam to be cone shaped. ranging between 0 en 360 degrees,
This is helpfull in the case te user wants to simulate a led with a smaller tan 360 degree viewing angle. with a high number of rays being effective, resulting in a more dense ray beam without excessive calulation time.