Diffraction grating support (#36)
Additions are the support for simple diffraction gratings and the order parameter of rays, needed to define the order of diffraction.
I updated the glass material catalogue with some new materials. The getMaterial function is used by the grating class aswell, so it is now as a standalone function.
A bug was removed happening during total reflection wich would assign the index of refraction of the 2nd material to the last indices of refraction (if total reflection on lens-air surface this would be 1, which is ok, but if on lens-lens surface this would be n of the second lens, which would mess up the "last index of refraction chain"). As the last index of refaction does not change during total reflection, nothing new should be applied to this list, i think.
There is now an Icon for diffraction gratings.