Modify mirror to transmit and reflect for testing
Before implementing beamsplitters properly, use the pre-existing mirror element
to test the necessary code changes.

Instead of a single 'dNewRay', the code now uses a list of of those.

The changes already work turning a mirror object into a beamsplitter: if making
a plane a mirror, a ray is reflected and another is transmitted.

If the a cube is used instead of a plane, the transmitted rays will be reflected
and transmitted when exiting the cube. This might be desired for tracking
residual reflection but there is not reflection coefficient yet. Hence, when
trying to use, e.g., a thorlabs beamsplitter step file, the user should add a
face right where the actual coating is (the face where the two triangular halves
of the cube are glued together).

Turning the cube into a lens (piece of glass) with the mirror surface inside the
cube, however, does not work: the beam is only reflected but not transmitted
anymore. Look into this with low priority (in most of my models the angle of
incidence will be normal anyway).