Add new optical element: splitter
All necessary classes were added (copied and adjusted from the mirror classes).
The new icon is based on the mirror icon and was modifed via Claude AI to
include a transmitted beam.

The splitter works as described in the last commit (only single faces/planes
should be used as splitters).

There are currently problems when trying to use the optics workbench in an
assembly4 assembly: added mirrors do not seem to work. At least when the linked
parts are assigned to be mirrors after insertion. I believe this worked in the
past, but maybe I added the mirrors in files prior to inserting them in the
assembly.